requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
pymysql
pydantic
python-jose
passlib
bcrypt
python-multipart